updated to newest tasks, mine were apperently outdated

+++ 02ex.c
@@ -17,7 +17,8 @@ und Ihre Lösungen zwischen den "BEGIN CHANGES" und "END CHANGES" eintragen.
 
 #include <stdio.h>
 
-int sum_divisible_both(int n) {
+int sum_divisible_both(int n)
+{
     int result = 0;
     /*
     Aufgabe 1a:
@@ -40,7 +41,8 @@ int sum_divisible_both(int n) {
     return result;
 }
 
-int sum_divisible_or(int n) {
+int sum_divisible_or(int n)
+{
     int result = 0;
     /*
     Aufgabe 1b:
@@ -57,7 +59,8 @@ int sum_divisible_or(int n) {
     return result;
 }
 
-int sum_divisible_either(int n) {
+int sum_divisible_either(int n)
+{
     int result = 0;
     /*
     Aufgabe 1c:
@@ -80,8 +83,9 @@ int sum_divisible_either(int n) {
     return result;
 }
 
-int signum(int x) {
-    long int result = 0;
+int signum(int x)
+{
+    int result = 0;
     /*
     Aufgabe 2a:
     Weisen Sie der Variable `result` den Wert `-1` zu, falls `x` strikt negativ ist, `0` falls `x` gleich `0` ist, und
@@ -99,7 +103,8 @@ int signum(int x) {
     return result;
 }
 
-int is_leapyear(int year) {
+int is_leapyear(int year)
+{
     /*
     Aufgabe 2b:
     Im gregorianischen Kalender ist ein Jahr ein Schaltjahr, genau dann wenn die Jahreszahl durch 400 teilbar ist, oder
@@ -118,7 +123,8 @@ int is_leapyear(int year) {
     /* END CHANGES */
 }
 
-int greatest_of_three(int a, int b, int c) {
+int greatest_of_three(int a, int b, int c)
+{
     int result = 0;
     /*
     Aufgabe 2c:
@@ -131,7 +137,8 @@ int greatest_of_three(int a, int b, int c) {
     return result;
 }
 
-int is_prime(int n) {
+int is_prime(int n)
+{
     int result = 1;
     /*
     Aufgabe 3a:
@@ -153,7 +160,8 @@ int is_prime(int n) {
     return result;
 }
 
-int sum_first_primes(int n) {
+int sum_first_primes(int n)
+{
     int result = 0;
     /*
     Aufgabe 3b:
